+++ src/commonMain/kotlin/com/github/insanusmokrassar/TelegramBotAPI/types/message/content/media/AnimationContent.kt
@@ -5,8 +5,10 @@ import com.github.insanusmokrassar.TelegramBotAPI.CommonAbstracts.TextPart
 import com.github.insanusmokrassar.TelegramBotAPI.requests.abstracts.Request
 import com.github.insanusmokrassar.TelegramBotAPI.requests.send.media.SendAnimation
 import com.github.insanusmokrassar.TelegramBotAPI.types.ChatIdentifier
+import com.github.insanusmokrassar.TelegramBotAPI.types.InputMedia.InputMedia
+import com.github.insanusmokrassar.TelegramBotAPI.types.InputMedia.InputMediaAnimation
 import com.github.insanusmokrassar.TelegramBotAPI.types.MessageIdentifier
-import com.github.insanusmokrassar.TelegramBotAPI.types.ParseMode.HTMLParseMode
+import com.github.insanusmokrassar.TelegramBotAPI.types.ParseMode.*
 import com.github.insanusmokrassar.TelegramBotAPI.types.buttons.KeyboardMarkup
 import com.github.insanusmokrassar.TelegramBotAPI.types.files.AnimationFile
 import com.github.insanusmokrassar.TelegramBotAPI.types.files.DocumentFile
@@ -14,6 +16,7 @@ import com.github.insanusmokrassar.TelegramBotAPI.types.message.abstracts.Conten
 import com.github.insanusmokrassar.TelegramBotAPI.types.message.abstracts.Message
 import com.github.insanusmokrassar.TelegramBotAPI.types.message.content.abstracts.MediaContent
 import com.github.insanusmokrassar.TelegramBotAPI.utils.toHtmlCaptions
+import com.github.insanusmokrassar.TelegramBotAPI.utils.toMarkdownV2Captions
 
 data class AnimationContent(
     override val media: AnimationFile,
@@ -38,5 +41,15 @@ data class AnimationContent(
         disableNotification,
         replyToMessageId,
         replyMarkup
+    )
+
+    override fun asInputMedia(): InputMediaAnimation = InputMediaAnimation(
+        media.fileId,
+        toMarkdownV2Captions().first(),
+        MarkdownV2,
+        media.width,
+        media.height,
+        media.duration,
+        media.thumb ?.fileId
     )
 }