+++ src/commonMain/kotlin/com/github/insanusmokrassar/TelegramBotAPI/types/message/content/media/DocumentContent.kt
@@ -5,14 +5,18 @@ import com.github.insanusmokrassar.TelegramBotAPI.CommonAbstracts.TextPart
 import com.github.insanusmokrassar.TelegramBotAPI.requests.abstracts.Request
 import com.github.insanusmokrassar.TelegramBotAPI.requests.send.media.SendDocument
 import com.github.insanusmokrassar.TelegramBotAPI.types.ChatIdentifier
+import com.github.insanusmokrassar.TelegramBotAPI.types.InputMedia.InputMedia
+import com.github.insanusmokrassar.TelegramBotAPI.types.InputMedia.InputMediaDocument
 import com.github.insanusmokrassar.TelegramBotAPI.types.MessageIdentifier
 import com.github.insanusmokrassar.TelegramBotAPI.types.ParseMode.HTMLParseMode
+import com.github.insanusmokrassar.TelegramBotAPI.types.ParseMode.MarkdownV2
 import com.github.insanusmokrassar.TelegramBotAPI.types.buttons.KeyboardMarkup
 import com.github.insanusmokrassar.TelegramBotAPI.types.files.DocumentFile
 import com.github.insanusmokrassar.TelegramBotAPI.types.message.abstracts.ContentMessage
 import com.github.insanusmokrassar.TelegramBotAPI.types.message.abstracts.Message
 import com.github.insanusmokrassar.TelegramBotAPI.types.message.content.abstracts.MediaContent
 import com.github.insanusmokrassar.TelegramBotAPI.utils.toHtmlCaptions
+import com.github.insanusmokrassar.TelegramBotAPI.utils.toMarkdownV2Captions
 
 data class DocumentContent(
     override val media: DocumentFile,
@@ -33,5 +37,12 @@ data class DocumentContent(
         disableNotification,
         replyToMessageId,
         replyMarkup
+    )
+
+    override fun asInputMedia(): InputMediaDocument = InputMediaDocument(
+        media.fileId,
+        toMarkdownV2Captions().first(),
+        MarkdownV2,
+        media.thumb ?.fileId
     )
 }