+++ src/commonMain/kotlin/com/github/insanusmokrassar/TelegramBotAPI/types/message/content/media/PhotoContent.kt
@@ -4,10 +4,10 @@ import com.github.insanusmokrassar.TelegramBotAPI.CommonAbstracts.TextPart
 import com.github.insanusmokrassar.TelegramBotAPI.requests.abstracts.Request
 import com.github.insanusmokrassar.TelegramBotAPI.requests.send.media.SendPhoto
 import com.github.insanusmokrassar.TelegramBotAPI.types.ChatIdentifier
-import com.github.insanusmokrassar.TelegramBotAPI.types.InputMedia.InputMediaPhoto
-import com.github.insanusmokrassar.TelegramBotAPI.types.InputMedia.MediaGroupMemberInputMedia
+import com.github.insanusmokrassar.TelegramBotAPI.types.InputMedia.*
 import com.github.insanusmokrassar.TelegramBotAPI.types.MessageIdentifier
 import com.github.insanusmokrassar.TelegramBotAPI.types.ParseMode.HTMLParseMode
+import com.github.insanusmokrassar.TelegramBotAPI.types.ParseMode.MarkdownV2
 import com.github.insanusmokrassar.TelegramBotAPI.types.buttons.KeyboardMarkup
 import com.github.insanusmokrassar.TelegramBotAPI.types.files.PhotoSize
 import com.github.insanusmokrassar.TelegramBotAPI.types.files.biggest
@@ -16,6 +16,7 @@ import com.github.insanusmokrassar.TelegramBotAPI.types.message.abstracts.Messag
 import com.github.insanusmokrassar.TelegramBotAPI.types.message.content.abstracts.MediaCollectionContent
 import com.github.insanusmokrassar.TelegramBotAPI.types.message.content.abstracts.MediaGroupContent
 import com.github.insanusmokrassar.TelegramBotAPI.utils.toHtmlCaptions
+import com.github.insanusmokrassar.TelegramBotAPI.utils.toMarkdownV2Captions
 
 data class PhotoContent(
     override val mediaCollection: List<PhotoSize>,
@@ -43,5 +44,11 @@ data class PhotoContent(
         media.fileId,
         toHtmlCaptions().firstOrNull(),
         HTMLParseMode
+    )
+
+    override fun asInputMedia(): InputMediaPhoto = InputMediaPhoto(
+        media.fileId,
+        toMarkdownV2Captions().first(),
+        MarkdownV2
     )
 }